src/requirements.txt
# Dependencies - requires Apple Silicon (M1/M2/M3/M4) for training
typer[all]>=0.12.0        # CLI framework
pyyaml>=6.0
tqdm>=4.60
datasets>=2.14.0          # HuggingFace datasets (golang_datasets in config.yaml)
huggingface_hub[cli]>=0.23.0
mlx>=0.22.0               # Apple Silicon only
mlx-lm>=0.21.0            # Apple Silicon only
gguf>=0.18.0              # GGUF export
torch                     # required by llama.cpp convert_hf_to_gguf.py
